refactor: modularize API calls and state management in SearchPage
- Separated API calls for fetching categories and meals into custom hooks (useFetchCategories, useFetchMeals).
- Improved readability of handleToggleSave by replacing ternary with if statement, ensuring ESLint compliance.
- Simplified category and meal retrieval with hooks to centralize loading and error handling.
- Created CategoryCard component for modular category display.
- Enhanced documentation on complex logic for improved maintainability.
